APMRover2: correct some style

+++ APMrover2/Parameters.cpp
@@ -599,15 +599,15 @@ void Rover::load_parameters(void)
         cliSerial->printf("done.\n");
     }
 
-    unsigned long before = micros();
+    uint32_t before = micros();
     // Load all auto-loaded EEPROM variables
     AP_Param::load_all();
 
     AP_Param::set_frame_type_flags(AP_PARAM_FRAME_ROVER);
-    
+
     SRV_Channels::set_default_function(CH_1, SRV_Channel::k_steering);
     SRV_Channels::set_default_function(CH_3, SRV_Channel::k_throttle);
-    
+
     const uint8_t old_rc_keys[14] = { Parameters::k_param_rc_1_old,  Parameters::k_param_rc_2_old,
                                       Parameters::k_param_rc_3_old,  Parameters::k_param_rc_4_old,
                                       Parameters::k_param_rc_5_old,  Parameters::k_param_rc_6_old,
